+++ packages/yourloops/components/dashboard-widgets/patient-dashboard.tsx
@@ -29,10 +29,9 @@ import React, { type FunctionComponent } from 'react'
 import { type Patient } from '../../lib/patient/models/patient.model'
 import DeviceUsage from 'blip/app/components/chart/deviceUsage.js'
 import type MedicalDataService from 'medical-domain'
-import { DatumType, type TimePrefs } from 'medical-domain'
+import { type DateFilter, DatumType, type MedicalData, type TimePrefs } from 'medical-domain'
 import { type BgPrefs } from 'dumb'
-import type DataUtil from 'tidepool-viz/src/utils/data'
-import moment from 'moment-timezone'
+import DataUtil from 'tidepool-viz/src/utils/data'
 import Grid from '@mui/material/Grid'
 import AccessTime from '@mui/icons-material/AccessTime'
 import { useTranslation } from 'react-i18next'
@@ -52,15 +51,13 @@ import { makeStyles } from 'tss-react/mui'
 import ChatWidget from '../chat/chat-widget'
 import { useSelectedTeamContext } from '../../lib/selected-team/selected-team.provider'
 import { PRIVATE_TEAM_ID, useTeam } from '../../lib/team/team.hook'
+import { DEFAULT_DASHBOARD_TIME_RANGE_DAYS } from '../patient-data/patient-data.utils'
 
 interface PatientDashboardProps {
   bgPrefs: BgPrefs
-  dashboardEpochDate: number
-  dataUtil: typeof DataUtil
   goToDailySpecificDate: (date: number | Date) => void
   loading: boolean
   medicalDataService: MedicalDataService
-  msRange: number
   patient: Patient
   timePrefs: TimePrefs
 }
@@ -76,12 +73,9 @@ const useStyle = makeStyles()((theme) => ({
 export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props) => {
   const {
     bgPrefs,
-    dashboardEpochDate,
-    dataUtil,
     goToDailySpecificDate,
     loading,
     medicalDataService,
-    msRange,
     patient,
     timePrefs
   } = props
@@ -93,14 +87,33 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
   const { classes, theme } = useStyle()
 
   const isMobileBreakpoint: boolean = useMediaQuery(theme.breakpoints.only('xs'))
-  const endpoints = [
-    moment.utc(dashboardEpochDate - msRange).toISOString(), // start
-    moment.utc(dashboardEpochDate).toISOString() // end
-  ]
-  const dateFilter = {
-    start: dashboardEpochDate - msRange,
-    end: dashboardEpochDate
+
+  const computeDateFilter = (data: MedicalData): DateFilter => {
+    const now = new Date()
+    const dashboardStartDate = new Date()
+    dashboardStartDate.setDate(now.getDate() - DEFAULT_DASHBOARD_TIME_RANGE_DAYS)
+    const dateRangeData = [...data.smbg, ...data.cbg]
+    const localDates = dateRangeData
+      .filter(bgData => new Date(bgData.normalTime) >= dashboardStartDate)
+      .map((bgData) => bgData.normalTime)
+      .sort((a, b) => a.localeCompare(b))
+    const dateRangeSet = new Set(localDates)
+    const dataRangeArray = Array.from(dateRangeSet)
+    return {
+      start: new Date(dataRangeArray[0]).valueOf(),
+      end: new Date(dataRangeArray.pop()).valueOf()
+    }
   }
+
+  const dateFilter = computeDateFilter(medicalData)
+  const endpoints = [new Date(dateFilter.start).toJSON(), new Date(dateFilter.end).toJSON()]
+
+  const dataUtil = new DataUtil(medicalDataService.data, {
+    bgPrefs,
+    timePrefs,
+    endpoints
+  })
+
   const isSelectedTeamPrivate = selectedTeam?.id === PRIVATE_TEAM_ID
   const isCaregiver = user.isUserCaregiver()
   const isPatientWithNoTeams = user.isUserPatient() && getMedicalTeams().length === 0
@@ -144,7 +157,6 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
         <PatientStatisticsWidget
           medicalData={medicalData}
           bgPrefs={bgPrefs}
-          bgType={dataUtil.bgSource}
           dateFilter={dateFilter}
         >
           <Stats