+++ CHANGELOG.md
@@ -134,6 +134,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2522 PDF/CSV report generation does not take into account data that was not loaded in daily
 - YLP-2524 Override information is hardly visible
 - YLP-2529 Fix median in trends view
+- YLP-2535 Dashboard statistics should only be computed on days having data
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team